Put main tests classes in seperate files
