+++ backend/tests/base.py
@@ -1,0 +1,42 @@
+from fastapi.testclient import TestClient
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
+import pytest
+
+from catalog_app.main import app, ROOT_PATH
+from catalog_app.database import Base
+from catalog_app.dependencies import get_db
+
+SQLALCHEMY_DATABASE_URL = "sqlite:///./test.db"
+
+engine = create_engine(
+    SQLALCHEMY_DATABASE_URL, connect_args={"check_same_thread": False}
+)
+TestingSessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+
+
+Base.metadata.create_all(bind=engine)
+def overwrite_get_db():
+    try:
+        db = TestingSessionLocal()
+        yield db
+    finally:
+        db.close()
+
+app.dependency_overrides[get_db] = overwrite_get_db
+
+@pytest.fixture(autouse=True)
+def cleanup_db():
+    Base.metadata.drop_all(bind=engine)
+    Base.metadata.create_all(bind=engine)
+    yield
+    Base.metadata.drop_all(bind=engine)
+
+@pytest.fixture
+def session():
+    for session in overwrite_get_db():
+        return session
+
+@pytest.fixture
+def client():
+    return TestClient(app)
